feat(bindings): moved franka exceptions into own submodule
