+++ python/rcsss/_core/hw/exceptions.pyi
@@ -1,0 +1,44 @@
+# ATTENTION: auto generated from C++ code, use `make stubgen` to update!
+"""
+exceptions from the hardware module
+"""
+from __future__ import annotations
+
+__all__ = [
+    "FrankaCommandException",
+    "FrankaControlException",
+    "FrankaException",
+    "FrankaIncompatibleVersionException",
+    "FrankaInvalidOperationException",
+    "FrankaModelException",
+    "FrankaNetworkException",
+    "FrankaProtocolException",
+    "FrankaRealtimeException",
+]
+
+class FrankaCommandException(RuntimeError):
+    pass
+
+class FrankaControlException(RuntimeError):
+    pass
+
+class FrankaException(RuntimeError):
+    pass
+
+class FrankaIncompatibleVersionException(RuntimeError):
+    pass
+
+class FrankaInvalidOperationException(RuntimeError):
+    pass
+
+class FrankaModelException(RuntimeError):
+    pass
+
+class FrankaNetworkException(RuntimeError):
+    pass
+
+class FrankaProtocolException(RuntimeError):
+    pass
+
+class FrankaRealtimeException(RuntimeError):
+    pass
