+++ src/pybind/rcsss.cpp
@@ -312,6 +312,27 @@ PYBIND11_MODULE(_core, m) {
       .def("get_state", &rcs::hw::FrankaHand::get_state)
       .def("homing", &rcs::hw::FrankaHand::homing);
 
+  auto hw_except =
+      hw.def_submodule("exceptions", "exceptions from the hardware module");
+  py::register_exception<franka::Exception>(hw_except, "FrankaException",
+                                            PyExc_RuntimeError);
+  py::register_exception<franka::ModelException>(
+      hw_except, "FrankaModelException", PyExc_RuntimeError);
+  py::register_exception<franka::NetworkException>(
+      hw_except, "FrankaNetworkException", PyExc_RuntimeError);
+  py::register_exception<franka::ProtocolException>(
+      hw_except, "FrankaProtocolException", PyExc_RuntimeError);
+  py::register_exception<franka::IncompatibleVersionException>(
+      hw_except, "FrankaIncompatibleVersionException", PyExc_RuntimeError);
+  py::register_exception<franka::ControlException>(
+      hw_except, "FrankaControlException", PyExc_RuntimeError);
+  py::register_exception<franka::CommandException>(
+      hw_except, "FrankaCommandException", PyExc_RuntimeError);
+  py::register_exception<franka::RealtimeException>(
+      hw_except, "FrankaRealtimeException", PyExc_RuntimeError);
+  py::register_exception<franka::InvalidOperationException>(
+      hw_except, "FrankaInvalidOperationException", PyExc_RuntimeError);
+
   // SIM MODULE
   auto sim = m.def_submodule("sim", "sim module");
   py::class_<rcs::sim::FR3Config, rcs::common::RConfig>(sim, "FR3Config")
@@ -333,23 +354,4 @@ PYBIND11_MODULE(_core, m) {
       .def("get_state", &rcs::sim::FR3::get_state)
       .def("reset", &rcs::sim::FR3::reset)
       .def("clear_markers", &rcs::sim::FR3::clear_markers);
-
-  py::register_exception<franka::Exception>(m, "FrankaException",
-                                            PyExc_RuntimeError);
-  py::register_exception<franka::ModelException>(m, "FrankaModelException",
-                                                 PyExc_RuntimeError);
-  py::register_exception<franka::NetworkException>(m, "FrankaNetworkException",
-                                                   PyExc_RuntimeError);
-  py::register_exception<franka::ProtocolException>(
-      m, "FrankaProtocolException", PyExc_RuntimeError);
-  py::register_exception<franka::IncompatibleVersionException>(
-      m, "FrankaIncompatibleVersionException", PyExc_RuntimeError);
-  py::register_exception<franka::ControlException>(m, "FrankaControlException",
-                                                   PyExc_RuntimeError);
-  py::register_exception<franka::CommandException>(m, "FrankaCommandException",
-                                                   PyExc_RuntimeError);
-  py::register_exception<franka::RealtimeException>(
-      m, "FrankaRealtimeException", PyExc_RuntimeError);
-  py::register_exception<franka::InvalidOperationException>(
-      m, "FrankaInvalidOperationException", PyExc_RuntimeError);
 }
